BoardCheckers

Description

This module exports all the functions required by playCheckers function in Checkers game.

Documentation

data Square

This data type represents a Square A Square consists of two ints

Constructors

Square Int Int

data Player

This data type represents a Player

Constructors

Red
To represent Red Player

Black
To represent Black Player

data PieceType

This data type represents a PieceType PieceType can be either Pawn or King

Constructors

Pawn
To represent Pawn

King
To represent King

data Piece

This data type represents Piece Piece can either be empty or fully qualified by its PieceType and Player

Constructors

Empty
To represent Empty Piece, Empty pieces are not shown on board

Piece Player PieceType
To represent a Piece denoted by its PieceType and Player

data Move

This data type represents a Move Can be either a Jump or a March

Constructors

Jump Square Square
To represent Jump from Square to Square

March Square Square
To represent March from Square to Square

isJump :: Move -> Bool

Take a Move and decide if it is a Jump or not

getDestination :: Move -> Square

Take a Move and return its destination Square

getSource :: Move -> Square

Take a Move and return its source Square

data BoardMap

This data type represents BoardMap

Constructors

BoardMap (Map Square Piece)
BoardMap is a Map from Square to Piece

class Board board where

Class Board

Minimal complete definition

seekBoard, removePiece, addPiece, replacePiece, validMoves, playMove

Methods

seekBoard :: board -> Square -> Piece

Return the Piece that is at the Square in the board

validMoves :: board -> Player -> [Move]

Return the valid moves of the specified Player

playMove :: board -> Move -> board

Play the Move on the board and return the new board

data Game

This data type represents Game

Constructors

Game BoardMap [Move] Player Square
A Game consists of the BoardMap, possible moves of the current Player, the current Player and the Square which was clicked

mValidJumps :: BoardMap -> (Square, Piece) -> [Move]

Return the valid jumps from the passed Square of the passed Player

startGame :: BoardMap

Return the initial BoardMap