BoardGo

Description

This module exports all the functions required by playGo function in Go game.

Documentation

data Point

This data type represents a Point in Board of Go (i.e. intersection of two lines)

Constructors

Point Int Int
Point constructor takes two coordinates x and y

data Stone

This data type represents Stone placed on the board

Constructors

Black
To represent Black stone

White
To represent White stone

Ko
Ko is not shown on the board. Placed when Ko situation occurs

Empty
To represent that nothing is placed on the point

getOppositeStone :: Stone -> Stone

getOppositeStone returns opposite stone of given Stone

data Move

This data type represents a Move that can be performed on the board by a player

Constructors

Pass Stone
Player with this Stone plays pass

Move Point Stone
Player with this Stone plays on a Point

getPiece :: Move -> Stone

getPiece returns the Stone from a given Move

data GameStatus

GameStatus represents state of the game if game is alive or dead or over

Constructors

Alive
game is running

Dead
game is over and hopeless string is counted

Over
Game over and results are shown on the screen

data Game

This data type reprsents Game

Constructors

Game (Map Point Stone) Move Int Int Int GameStatus

createGame :: Int -> Game

createGame takes a size and returns a Game object

killGame :: Game -> Game

killGame changes game status from Alive to Dead and increases the score of Black as White plays the last Pass.

seekBoard :: Game -> Point -> Stone

seekBoard returns the Stone at given Point in the Game

removeKo :: Game -> Game

removeKo removes Ko from the Game

removePiece :: Map Point Stone -> Point -> Map Point Stone

removePiece removes the stone from given point in the map

playPass :: Game -> Stone -> Game

playPass adds Pass move by the Stone on the Game

validMove :: Game -> Point -> Stone -> Bool

validMove checks if the move of given Point to given Stone is valid or not

playMove :: Game -> Point -> Stone -> Game

playMove adds Move to Point by Stone on the Game. After playing the move, also removes the captured points of the opposite stone if present

finishGame :: Game -> Game

finishGame finds all the captured territories in the Game and updates the correspoing scores of the Stone

getWinner :: Game -> String

getWinner declares the winner depending on the scores of the stones