Go

Description

This module has only one function playGo that asks user for size as input, creates a game and starts the game window.

Documentation

playGo :: IO ()

PlayGo functions asks for size of board, creates a Game object and starts the game display